fixed: update for mavlink

+++ libraries/GCS_MAVLink/GCS.h
@@ -219,7 +219,7 @@ public:
     void send_local_position_ned(float px, float py, float pz, float vx, float vy, float vz) const;
     bool locked() const;
 
-    void send_cndn_trigger(double lat, double lng);
+    void send_cndn_trigger(uint32_t lat, uint32_t lng);
     void send_cndn_request(uint8_t sess, uint16_t size, uint16_t offset);
 
     // return a bitmap of active channels. Used by libraries to loop
@@ -880,7 +880,7 @@ public:
     void send_command_long(uint32_t msgid, uint8_t confirmation=0, float p1=0.0f, float p2=0.0f, float p3=0.0f, float p4=0.0f, float p5=0.0f, float p6=0.0f, float p7=0.0f);
     void send_local_position_ned(float px=0.0f, float py=0.0f, float pz=0.0f, float vx=0.0f, float vy=0.0f, float vz=0.0f);
 
-    void send_cndn_trigger(double lat, double lng);
+    void send_cndn_trigger(uint32_t lat, uint32_t lng);
     void send_cndn_request(uint8_t sess, uint16_t size, uint16_t offset);
 
 protected:
