+++ libraries/GCS_MAVLink/GCS_Common.cpp
@@ -4713,13 +4713,13 @@ void GCS::send_local_position_ned(float px, float py, float pz, float vx, float 
     }
 }
 
-void GCS::send_cndn_trigger(double lat, double lng) {
+void GCS::send_cndn_trigger(uint32_t lat, uint32_t lng) {
     for (uint8_t i=0; i<num_gcs(); i++) {
         chan(i)->send_cndn_trigger(lat, lng);
     }
 }
 
-void GCS_MAVLINK::send_cndn_trigger(double lat, double lng) {
+void GCS_MAVLINK::send_cndn_trigger(uint32_t lat, uint32_t lng) {
     if (HAVE_PAYLOAD_SPACE(chan, CNDN_TRIGGER)) {
         mavlink_msg_cndn_trigger_send(chan, AP_HAL::millis(), lat, lng);
     }
